rename Navigation to WaypointNavigation
