+++ field_friend/automations/navigation/__init__.py
@@ -1,15 +1,15 @@
 from .field_navigation import FieldNavigation, RowSegment
 from .implement_demo_navigation import ImplementDemoNavigation
-from .navigation import Navigation, PathSegment, WorkflowException, WorkingSegment
 from .straight_line_navigation import StraightLineNavigation
+from .waypoint_navigation import PathSegment, WaypointNavigation, WorkflowException, WorkingSegment
 
 __all__ = [
     'FieldNavigation',
     'ImplementDemoNavigation',
-    'Navigation',
     'PathSegment',
     'RowSegment',
     'StraightLineNavigation',
+    'WaypointNavigation',
     'WorkflowException',
     'WorkingSegment',
 ]
