+++ tests/test_navigation.py
@@ -14,7 +14,7 @@ from rosys.testing import assert_point, forward
 from field_friend import System
 from field_friend.automations import AutomationWatcher
 from field_friend.automations.implements import Implement, Recorder
-from field_friend.automations.navigation import Navigation, PathSegment, StraightLineNavigation
+from field_friend.automations.navigation import PathSegment, StraightLineNavigation, WaypointNavigation
 from field_friend.hardware.double_wheels import WheelsSimulationWithAcceleration
 
 
@@ -65,7 +65,7 @@ async def test_driving_to_exact_positions(system: System):
         assert system.robot_locator.pose.point.y == pytest.approx(
             stopper_implement.current_target_position.y, abs=0.001)
         await forward(0.1)  # give robot time to update position
-    system.current_navigation.linear_speed_limit = Navigation.LINEAR_SPEED_LIMIT
+    system.current_navigation.linear_speed_limit = WaypointNavigation.LINEAR_SPEED_LIMIT
     await forward(until=lambda: system.automator.is_stopped)
     assert system.robot_locator.pose.x == pytest.approx(system.current_navigation.length, abs=0.001)
